requirements.txt
Flask
mysql-connector-python
gunicorn
twilio
